Cleanup nq release code
